Dropbox: changed warning lavel to w4
git-svn-id: http://svn.miranda-ng.org/main/trunk@11419 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Dropbox/src/dropbox.h
@@ -45,7 +45,7 @@ private:
 	ULONG  hMessageProcess;
 
 	HANDLE hFileSentEventHook;
-	
+
 	MCONTACT hDefaultContact;
 	MCONTACT hTransferContact;
 
@@ -95,8 +95,8 @@ private:
 
 	// transrers
 	int SendFile(const char *fileName, const char *data, int length);
-	int SendFileChunkedFirst(const char *data, int length, char *uploadId, int &offset);
-	int SendFileChunkedNext(const char *data, int length, const char *uploadId, int &offset);
+	int SendFileChunkedFirst(const char *data, int length, char *uploadId, size_t &offset);
+	int SendFileChunkedNext(const char *data, int length, const char *uploadId, size_t &offset);
 	int SendFileChunkedLast(const char *fileName, const char *uploadId);
 
 	int CreateFolder(const char *folderName);
